Extract the hovered/clicked logic from SimpleUI into a core UIEngine class
