+++ Sprocket/Sprocket.h
@@ -16,6 +16,7 @@
 
 #include "UI/DevUI.h"
 #include "UI/SimpleUI.h"
+#include "UI/UIEngine.h"
 
 // EVENTS
 #include "Events/Event.h"
